move duplicates manager to simple_reader/
