--- ydb/core/tx/columnshard/engines/reader/duplicates/events.cpp
@@ -1,3 +1,0 @@
-#include "events.h"
-
-namespace NKikimr::NOlap::NReader {}
