--- ydb/core/tx/columnshard/engines/reader/duplicates/fetching.cpp
@@ -1,4 +1,0 @@
-#include "fetching.h"
-
-namespace NKikimr::NOlap::NReader {
-}
