--- ydb/core/tx/columnshard/engines/reader/duplicates/subscriber.cpp
@@ -1,3 +1,0 @@
-#include "subscriber.h"
-
-namespace NKikimr::NOlap::NReader {}
