--- ydb/core/tx/columnshard/engines/reader/duplicates/ya.make
@@ -1,19 +1,0 @@
-RECURSE_FOR_TESTS(
-    ut
-)
-
-LIBRARY()
-
-SRCS(
-    manager.cpp
-    events.cpp
-    merge.cpp
-    subscriber.cpp
-    fetching.cpp
-)
-
-PEERDIR(
-    ydb/core/tx/columnshard/engines/reader/common_reader/iterator
-)
-
-END()
