+++ ydb/core/tx/columnshard/engines/reader/simple_reader/duplicates/events.h
@@ -3,30 +3,18 @@
 #include <ydb/core/formats/arrow/arrow_filter.h>
 #include <ydb/core/tx/columnshard/columnshard_private_events.h>
 #include <ydb/core/tx/columnshard/engines/reader/common_reader/iterator/source.h>
+#include <ydb/core/tx/columnshard/engines/reader/simple_reader/duplicates/subscriber.h>
 
 #include <ydb/library/actors/core/event_local.h>
 #include <ydb/library/conclusion/result.h>
 
 namespace NKikimr::NOlap::NReader::NSimple {
+
 class IDataSource;
-}
-
-namespace NKikimr::NOlap::NReader::NSimple::NDuplicateFiltering {
-
-class IFilterSubscriber {
-public:
-    virtual void OnFilterReady(NArrow::TColumnFilter&&) = 0;
-    virtual void OnFailure(const TString& reason) = 0;
-    virtual ~IFilterSubscriber() = default;
-};
 
 class TEvRequestFilter: public NActors::TEventLocal<TEvRequestFilter, NColumnShard::TEvPrivate::EvRequestFilter> {
 private:
-    NArrow::TSimpleRow MinPK;
-    NArrow::TSimpleRow MaxPK;
-    YDB_READONLY_DEF(ui64, SourceId);
-    YDB_READONLY_DEF(ui64, RecordsCount);
-    TSnapshot MaxVersion;
+    YDB_READONLY_DEF(std::shared_ptr<IDataSource>, Source);
     YDB_READONLY_DEF(std::shared_ptr<IFilterSubscriber>, Subscriber);
     YDB_READONLY_DEF(std::shared_ptr<const TAtomicCounter>, AbortionFlag);
