+++ ydb/core/tx/columnshard/engines/reader/simple_reader/duplicates/fetching.cpp
@@ -1,0 +1,3 @@
+#include "fetching.h"
+
+namespace NKikimr::NOlap::NReader::NSimple {}
