+++ ydb/core/tx/columnshard/engines/reader/simple_reader/duplicates/subscriber.cpp
@@ -1,0 +1,3 @@
+#include "subscriber.h"
+
+namespace NKikimr::NOlap::NReader::NSimple {}
