+++ ydb/core/tx/columnshard/engines/reader/simple_reader/duplicates/ut/ut_manager.cpp
@@ -1,15 +1,17 @@
-#include <ydb/core/tx/columnshard/engines/reader/duplicates/manager.h>
+#include <ydb/core/tx/columnshard/engines/reader/simple_reader/duplicates/manager.h>
+
 #include <library/cpp/testing/unittest/registar.h>
 
 using namespace NKikimr;
 using namespace NKikimr::NOlap;
 using namespace NKikimr::NOlap::NReader;
+using namespace NKikimr::NOlap::NReader::NSimple;
 
 Y_UNIT_TEST_SUITE(TIntervalCounter) {
     void TestDec(TIntervalCounter & counter, ui32 l, ui32 r, const std::vector<ui32>& expectedZeros) {
         const auto result = counter.DecAndGetZeros(l, r);
-        UNIT_ASSERT_EQUAL_C(
-            result, expectedZeros, TStringBuilder() << '[' << JoinSeq(',', result) << "] (actual) != [" << JoinSeq(',', expectedZeros) << "] (expected)");
+        UNIT_ASSERT_EQUAL_C(result, expectedZeros,
+            TStringBuilder() << '[' << JoinSeq(',', result) << "] (actual) != [" << JoinSeq(',', expectedZeros) << "] (expected)");
     }
 
     Y_UNIT_TEST(Count) {
