+++ ydb/core/tx/columnshard/engines/reader/simple_reader/duplicates/ya.make
@@ -1,3 +1,7 @@
+RECURSE_FOR_TESTS(
+    ut
+)
+
 LIBRARY()
 
 SRCS(
